Add documentation, support, and lint for tentative directories
Based on the discussion in the issue, we have landed on only allowing
directories named 'foo.tentative'. This PR documents this, adds support
for it in the only place that cares about tentative
(manifest/sourcefile.py), and adds a lint rule to enforce the naming.

Directories that break the lint rule are also renamed.

Fixes https://github.com/web-platform-tests/wpt/issues/23300