Make system color keywords compute to themselves
Prior to this change, system color keywords would be resolved to RGBA
values at computed-value time. With this change, we defer resolution
until resolved-value time, to align with css-color-4.

The kernel of the change is in StyleBuilderConverter which allows for
storing an unresolved system color keyword, and in StyleColor which can
now be constructed from a system color keyword and requires a
WebColorScheme value to resolve colors. The remainder of the change
consists of updates at call sites to plumb in the used color scheme.

Bug: 1081945
Change-Id: I3b704cdf38c72b5284ad61a20e28825ee3e8e9e8